agent_service/pyproject.toml
[tool.poetry]
name = "agent-service"
version = "0.1.0"
description = ""
authors = ["ce.liu <ce.liu@avepoint.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12.0"
fastapi = "^0.115.6"
langchain = "^0.3.9"
requests = "^2.32.3"
uvicorn = "^0.32.1"
langchain-openai = "^0.2.11"
langchain-community = "^0.3.9"
langchain-anthropic = "^0.3.0"
langchain-experimental = "^0.3.3"
matplotlib = "^3.9.3"
langchain-ollama = "^0.2.1"
chromadb = "^0.5.23"
autogen-agentchat = {extras = ["ollama"], version = "^0.2.39"}
langgraph = "^0.2.56"
markdownify = "^0.14.1"
pypdf = "^5.1.0"
flaml = {extras = ["automl"], version = "^2.3.2"}
sentence-transformers = "^3.3.1"
llama-index = "^0.12.8"
pyautogen = {extras = ["ollama"], version = "^0.6.0"}
llama-index-llms-ollama = "^0.5.0"
llama-index-embeddings-ollama = "^0.5.0"
tavily-python = "^0.5.0"
mammoth = "^1.8.0"
pdfminer = "^20191125"
pdfminer-six = "^20240706"
python-pptx = "^1.0.2"
puremagic = "^1.28"
pathvalidate = "^3.2.1"
playwright = "^1.49.1"
pygraphviz = "^1.14"
llama-index-llms-azure-openai = "^0.3.0"
pydantic-core = "^2.27.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
autogenstudio = "^0.1.5"

[isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
